refactor: split index.ts into config/ffmpeg-bootstrap/encoders/scan modules
