+++ index.ts
@@ -15,1065 +15,33 @@
 
 import path from "node:path";
 import { promises as fs } from "node:fs";
-import os from "node:os";
-import crypto from "node:crypto";
 
-type EncoderChoice = "cpu" | "qsv" | "nvenc" | "amf";
-
-interface BenchmarkConfig {
-  // Hardware acceleration availability
-  hasQsv: boolean;
-  hasNvenc: boolean;
-  hasAmf: boolean;
-
-  // Concurrency settings
-  chosenCpuConcurrency: number; // how many CPU encodes in parallel
-  chosenQsvConcurrency: number; // 0 or 1
-  chosenNvencConcurrency: number; // 0 or 1
-  chosenAmfConcurrency: number; // 0 or 1
-
-  // CPU encoding settings
-  cpuPreset: string; // libx264 preset
-  cpuCrf: number; // libx264 CRF
-  cpuVideoEncoder: string; // selected CPU encoder (libx264, h264_mf, libopenh264)
-  fallbackBitrate: string; // used if encoder doesn't support CRF (e.g. h264_mf, libopenh264)
-
-  // Audio settings
-  aacBitrate: string; // e.g. "160k"
-
-  // Hardware acceleration usage flags
-  useQsv: boolean; // whether to schedule QSV encodes
-  useNvenc: boolean; // whether to schedule NVENC encodes
-  useAmf: boolean; // whether to schedule AMF encodes
-
-  // Hardware-specific quality settings
-  qsvGlobalQuality: number; // h264_qsv global_quality (roughly like CRF)
-  nvencCq: number; // h264_nvenc constant quality
-  amfQpI: number; // h264_amf quantization parameter for I frames
-}
-
-interface ConvertOptions {
-  inputDir: string;
-  outputDir: string;
-  force: boolean;
-  dryRun: boolean;
-  configPath?: string;
-}
-
-const DEFAULT_CONFIG: BenchmarkConfig = {
-  // Hardware acceleration availability
-  hasQsv: false,
-  hasNvenc: false,
-  hasAmf: false,
-
-  // Concurrency settings
-  chosenCpuConcurrency: Math.max(
-    1,
-    Math.min(4, Math.floor((os.cpus()?.length ?? 4) / 3))
-  ),
-  chosenQsvConcurrency: 0,
-  chosenNvencConcurrency: 0,
-  chosenAmfConcurrency: 0,
-
-  // CPU encoding settings
-  cpuPreset: "veryfast",
-  cpuCrf: 20,
-  cpuVideoEncoder: "libx264",
-  fallbackBitrate: "3000k",
-
-  // Audio settings
-  aacBitrate: "160k",
-
-  // Hardware acceleration usage flags
-  useQsv: false,
-  useNvenc: false,
-  useAmf: false,
-
-  // Hardware-specific quality settings
-  qsvGlobalQuality: 23,
-  nvencCq: 20,
-  amfQpI: 20,
-};
-
-function getDefaultConfigPath(): string {
-  return resolvePath("tv-media-conv.config.json");
-}
-
-function resolvePath(p: string): string {
-  return path.isAbsolute(p) ? p : path.resolve(process.cwd(), p);
-}
-
-function getCacheBaseDir(): string {
-  if (process.platform === "win32") {
-    return (
-      process.env.LOCALAPPDATA ??
-      path.join(os.homedir(), "AppData", "Local")
-    );
-  }
-  if (process.platform === "darwin") {
-    return path.join(os.homedir(), "Library", "Caches");
-  }
-  return process.env.XDG_CACHE_HOME ?? path.join(os.homedir(), ".cache");
-}
-
-function getFfmpegCacheDir(): string {
-  return path.join(getCacheBaseDir(), "tv-media-conv", "ffmpeg");
-}
-
-async function resolveCommand(cmd: string): Promise<string | null> {
-  const candidates = process.platform === "win32" ? [cmd, `${cmd}.exe`] : [cmd];
-  for (const candidate of candidates) {
-    const resolved = await resolveWithWhere(candidate);
-    if (resolved) return resolved;
-  }
-  return null;
-}
-
-async function resolveWithWhere(cmd: string): Promise<string | null> {
-  const tool = process.platform === "win32" ? "where" : "which";
-  try {
-    const proc = Bun.spawn({
-      cmd: [tool, cmd],
-      stdout: "pipe",
-      stderr: "ignore",
-    });
-    const [code, out] = await Promise.all([proc.exited, proc.stdout?.text()]);
-    if (code !== 0) return null;
-    const first = (out ?? "")
-      .split(/\r?\n/)
-      .map((line) => line.trim())
-      .find(Boolean);
-    return first ?? null;
-  } catch {
-    return null;
-  }
-}
-
-async function canExecute(p: string): Promise<boolean> {
-  try {
-    await fs.access(p);
-    const test = Bun.spawn({
-      cmd: [p, "-version"],
-      stdout: "ignore",
-      stderr: "ignore",
-    });
-    const code = await test.exited;
-    return code === 0;
-  } catch {
-    return false;
-  }
-}
-
-function isPathInside(parent: string, child: string): boolean {
-  const relative = path.relative(parent, child);
-  return !!relative && !relative.startsWith("..") && !path.isAbsolute(relative);
-}
-
-function sha256Hex(buffer: ArrayBuffer): string {
-  return crypto
-    .createHash("sha256")
-    .update(Buffer.from(buffer))
-    .digest("hex");
-}
-
-async function fetchText(url: string): Promise<string> {
-  const response = await fetch(url);
-  if (!response.ok) {
-    throw new Error(`Failed to download ${url}: ${response.status} ${response.statusText}`);
-  }
-  return await response.text();
-}
-
-async function fetchArrayBuffer(url: string): Promise<ArrayBuffer> {
-  const response = await fetch(url);
-  if (!response.ok) {
-    throw new Error(`Failed to download ${url}: ${response.status} ${response.statusText}`);
-  }
-  return await response.arrayBuffer();
-}
-
-async function extractZipWindows(archivePath: string, extractDir: string): Promise<void> {
-  const escapedArchive = archivePath.replace(/'/g, "''");
-  const escapedExtract = extractDir.replace(/'/g, "''");
-  const command = `Expand-Archive -LiteralPath '${escapedArchive}' -DestinationPath '${escapedExtract}' -Force`;
-  const proc = Bun.spawn({
-    cmd: ["powershell", "-NoProfile", "-NonInteractive", "-Command", command],
-    stdout: "pipe",
-    stderr: "pipe",
-  });
-  const [code, err] = await Promise.all([proc.exited, proc.stderr?.text()]);
-  if (code !== 0) {
-    throw new Error(`Failed to extract ffmpeg zip: ${err ?? "unknown error"}`);
-  }
-}
-
-async function extractTarXz(archivePath: string, extractDir: string): Promise<void> {
-  const proc = Bun.spawn({
-    cmd: ["tar", "-xJf", archivePath, "-C", extractDir],
-    stdout: "pipe",
-    stderr: "pipe",
-  });
-  const [code, err] = await Promise.all([proc.exited, proc.stderr?.text()]);
-  if (code !== 0) {
-    throw new Error(
-      `Failed to extract ffmpeg archive. Ensure 'tar' and 'xz' are installed. ${err ?? ""}`
-    );
-  }
-}
-
-async function findExecutableInDir(
-  rootDir: string,
-  name: string
-): Promise<string | null> {
-  const exeName = process.platform === "win32" ? `${name}.exe` : name;
-  const walk = async (dir: string): Promise<string | null> => {
-    const entries = await fs.readdir(dir, { withFileTypes: true });
-    for (const entry of entries) {
-      const fullPath = path.join(dir, entry.name);
-      if (!isPathInside(rootDir, fullPath)) continue;
-      if (entry.isSymbolicLink()) continue;
-      if (entry.isDirectory()) {
-        const found = await walk(fullPath);
-        if (found) return found;
-      } else if (entry.isFile() && entry.name === exeName) {
-        return fullPath;
-      }
-    }
-    return null;
-  };
-  return walk(rootDir);
-}
-
-async function copyWindowsRuntimeFiles(fromDir: string, toDir: string): Promise<void> {
-  const entries = await fs.readdir(fromDir, { withFileTypes: true });
-  for (const entry of entries) {
-    if (!entry.isFile()) continue;
-    const srcPath = path.join(fromDir, entry.name);
-    const destPath = path.join(toDir, entry.name);
-    await fs.copyFile(srcPath, destPath);
-  }
-}
-
-async function ensureFfmpegTools(): Promise<{
-  ffmpeg: string;
-  ffprobe: string;
-}> {
-  const [ffmpegResolved, ffprobeResolved] = await Promise.all([
-    resolveCommand("ffmpeg"),
-    resolveCommand("ffprobe"),
-  ]);
-  if (
-    ffmpegResolved &&
-    ffprobeResolved &&
-    (await canExecute(ffmpegResolved)) &&
-    (await canExecute(ffprobeResolved))
-  ) {
-    return { ffmpeg: ffmpegResolved, ffprobe: ffprobeResolved };
-  }
-
-  const platform = process.platform;
-  const arch = process.arch;
-
-  const ffmpegDir = getFfmpegCacheDir();
-  const ffmpegPath = path.join(
-    ffmpegDir,
-    platform === "win32" ? "ffmpeg.exe" : "ffmpeg"
-  );
-  const ffprobePath = path.join(
-    ffmpegDir,
-    platform === "win32" ? "ffprobe.exe" : "ffprobe"
-  );
-
-  if (await canExecute(ffmpegPath) && await canExecute(ffprobePath)) {
-    return { ffmpeg: ffmpegPath, ffprobe: ffprobePath };
-  }
-
-  console.log("ffmpeg not found on PATH, downloading from BtbN/FFmpeg-Builds...");
-
-  let archiveName: string;
-  let archiveExt: string;
-  if (platform === "win32" && arch === "x64") {
-    archiveName = "ffmpeg-n8.1-latest-win64-gpl-8.1.zip";
-    archiveExt = ".zip";
-  } else if (platform === "linux" && arch === "x64") {
-    archiveName = "ffmpeg-n8.1-latest-linux64-gpl-8.1.tar.xz";
-    archiveExt = ".tar.xz";
-  } else if (platform === "linux" && arch === "arm64") {
-    archiveName = "ffmpeg-n8.1-latest-linuxarm64-gpl-8.1.tar.xz";
-    archiveExt = ".tar.xz";
-  } else {
-    throw new Error(
-      `ffmpeg not found on PATH and auto-download is not supported for ${platform}-${arch}. ` +
-        "Please install ffmpeg manually."
-    );
-  }
-
-  const releaseBase = "https://github.com/BtbN/FFmpeg-Builds/releases/download/latest";
-  const downloadUrl = `${releaseBase}/${archiveName}`;
-  const checksumUrl = `${releaseBase}/checksums.sha256`;
-
-  await fs.mkdir(ffmpegDir, { recursive: true });
-  const tempDir = path.join(ffmpegDir, "tmp");
-  const extractDir = path.join(tempDir, "extract");
-  const archivePath = path.join(tempDir, `ffmpeg${archiveExt}`);
-
-  let installed = false;
-  try {
-    await fs.rm(tempDir, { recursive: true, force: true });
-    await fs.mkdir(extractDir, { recursive: true });
-
-    const checksumsText = await fetchText(checksumUrl);
-    const checksumMatch = checksumsText
-      .split(/\r?\n/)
-      .map((line) => line.trim())
-      .filter(Boolean)
-      .map((line) => line.match(/^([a-f0-9]{64})\s+\*?(.+)$/i))
-      .find((match) => match && match[2] === archiveName);
-    if (!checksumMatch) {
-      throw new Error(`Checksum not found for ${archiveName} in checksums.sha256`);
-    }
-    const expectedChecksum = checksumMatch[1]!.toLowerCase();
-
-    console.log(`Downloading ffmpeg from ${downloadUrl}...`);
-    const archiveBuffer = await fetchArrayBuffer(downloadUrl);
-    const actualChecksum = sha256Hex(archiveBuffer);
-    if (actualChecksum !== expectedChecksum) {
-      throw new Error(
-        `Checksum mismatch for ${archiveName}. Expected ${expectedChecksum}, got ${actualChecksum}.`
-      );
-    }
-    await Bun.write(archivePath, archiveBuffer);
-
-    console.log("Extracting ffmpeg...");
-    if (archiveExt === ".zip") {
-      if (platform !== "win32") {
-        throw new Error("ZIP extraction is only supported on Windows for auto-download");
-      }
-      await extractZipWindows(archivePath, extractDir);
-    } else if (archiveExt === ".tar.xz") {
-      await extractTarXz(archivePath, extractDir);
-    }
-
-    const [foundFfmpeg, foundFfprobe] = await Promise.all([
-      findExecutableInDir(extractDir, "ffmpeg"),
-      findExecutableInDir(extractDir, "ffprobe"),
-    ]);
-
-    if (!foundFfmpeg || !foundFfprobe) {
-      throw new Error("Failed to find ffmpeg or ffprobe in downloaded archive");
-    }
-
-    await fs.copyFile(foundFfmpeg, ffmpegPath);
-    await fs.copyFile(foundFfprobe, ffprobePath);
-
-    if (platform === "win32") {
-      await copyWindowsRuntimeFiles(path.dirname(foundFfmpeg), ffmpegDir);
-    } else {
-      await fs.chmod(ffmpegPath, 0o755);
-      await fs.chmod(ffprobePath, 0o755);
-    }
-
-    if (!(await canExecute(ffmpegPath)) || !(await canExecute(ffprobePath))) {
-      throw new Error("Downloaded ffmpeg binaries failed to execute after install");
-    }
-
-    installed = true;
-  } catch (error) {
-    const message = error instanceof Error ? error.message : String(error);
-    throw new Error(`Failed to install ffmpeg automatically: ${message}`);
-  } finally {
-    await fs.rm(tempDir, { recursive: true, force: true });
-    if (!installed) {
-      await fs.rm(ffmpegDir, { recursive: true, force: true }).catch(() => {});
-    }
-  }
-
-  console.log(`ffmpeg downloaded and installed to ${ffmpegDir}`);
-  return { ffmpeg: ffmpegPath, ffprobe: ffprobePath };
-}
-
-async function getEncoders(ffmpegCmd: string): Promise<Set<string>> {
-  const proc = Bun.spawn({
-    cmd: [ffmpegCmd, "-hide_banner", "-encoders"],
-    stdout: "pipe",
-    stderr: "pipe",
-  });
-  const [code, out, err] = await Promise.all([
-    proc.exited,
-    proc.stdout?.text(),
-    proc.stderr?.text(),
-  ]);
-  const text = `${out ?? ""}\n${err ?? ""}`;
-  if (code !== 0) throw new Error("Failed to query encoders from ffmpeg");
-  const set = new Set<string>();
-  for (const line of text.split(/\r?\n/)) {
-    const m = line.match(
-      /\s([a-z0-9_]+)\s+\S+\s+H\.264/i
-    ) as RegExpMatchArray | null;
-    if (m && m[1]) set.add(m[1]);
-    const m2 = line.match(
-      /\s([a-z0-9_]+)\s+.*?AAC/i
-    ) as RegExpMatchArray | null;
-    if (m2 && m2[1]) set.add(m2[1]);
-  }
-  return set;
-}
-
-async function supportsQsv(ffmpegCmd: string): Promise<boolean> {
-  // Prefer direct encoder help query; ffmpeg returns error if encoder doesn't exist
-  try {
-    const helpProc = Bun.spawn({
-      cmd: [ffmpegCmd, "-hide_banner", "-v", "error", "-h", "encoder=h264_qsv"],
-      stdout: "ignore",
-      stderr: "pipe",
-    });
-    const [code, err] = await Promise.all([
-      helpProc.exited,
-      helpProc.stderr?.text(),
-    ]);
-    if (code === 0) return true;
-    const text = (err ?? "").toLowerCase();
-    if (text.includes("unknown encoder") || text.includes("not found"))
-      return false;
-  } catch {
-    // ignore and try fallback
-  }
-  // Fallback: scan encoders list
-  try {
-    const proc = Bun.spawn({
-      cmd: [ffmpegCmd, "-hide_banner", "-encoders"],
-      stdout: "pipe",
-      stderr: "pipe",
-    });
-    const [_, out, err] = await Promise.all([
-      proc.exited,
-      proc.stdout?.text(),
-      proc.stderr?.text(),
-    ]);
-    const text = `${out ?? ""}\n${err ?? ""}`.toLowerCase();
-    return text.includes("h264_qsv");
-  } catch {
-    return false;
-  }
-}
-
-async function supportsNvenc(ffmpegCmd: string): Promise<boolean> {
-  // Prefer direct encoder help query; ffmpeg returns error if encoder doesn't exist
-  try {
-    const helpProc = Bun.spawn({
-      cmd: [
-        ffmpegCmd,
-        "-hide_banner",
-        "-v",
-        "error",
-        "-h",
-        "encoder=h264_nvenc",
-      ],
-      stdout: "ignore",
-      stderr: "pipe",
-    });
-    const [code, err] = await Promise.all([
-      helpProc.exited,
-      helpProc.stderr?.text(),
-    ]);
-    if (code === 0) return true;
-    const text = (err ?? "").toLowerCase();
-    if (text.includes("unknown encoder") || text.includes("not found"))
-      return false;
-  } catch {
-    // ignore and try fallback
-  }
-  // Fallback: scan encoders list
-  try {
-    const proc = Bun.spawn({
-      cmd: [ffmpegCmd, "-hide_banner", "-encoders"],
-      stdout: "pipe",
-      stderr: "pipe",
-    });
-    const [_, out, err] = await Promise.all([
-      proc.exited,
-      proc.stdout?.text(),
-      proc.stderr?.text(),
-    ]);
-    const text = `${out ?? ""}
-${err ?? ""}`.toLowerCase();
-    return text.includes("h264_nvenc");
-  } catch {
-    return false;
-  }
-}
-
-async function supportsAmf(ffmpegCmd: string): Promise<boolean> {
-  // Prefer direct encoder help query; ffmpeg returns error if encoder doesn't exist
-  try {
-    const helpProc = Bun.spawn({
-      cmd: [ffmpegCmd, "-hide_banner", "-v", "error", "-h", "encoder=h264_amf"],
-      stdout: "ignore",
-      stderr: "pipe",
-    });
-    const [code, err] = await Promise.all([
-      helpProc.exited,
-      helpProc.stderr?.text(),
-    ]);
-    if (code === 0) return true;
-    const text = (err ?? "").toLowerCase();
-    if (text.includes("unknown encoder") || text.includes("not found"))
-      return false;
-  } catch {
-    // ignore and try fallback
-  }
-  // Fallback: scan encoders list
-  try {
-    const proc = Bun.spawn({
-      cmd: [ffmpegCmd, "-hide_banner", "-encoders"],
-      stdout: "pipe",
-      stderr: "pipe",
-    });
-    const [_, out, err] = await Promise.all([
-      proc.exited,
-      proc.stdout?.text(),
-      proc.stderr?.text(),
-    ]);
-    const text = `${out ?? ""}
-${err ?? ""}`.toLowerCase();
-    return text.includes("h264_amf");
-  } catch {
-    return false;
-  }
-}
-
-function parseSpeedFromFfmpegLog(stderrText: string): number | null {
-  // Look for the last occurrence of speed=1.23x
-  const matches = Array.from(
-    stderrText.matchAll(/speed\s*=\s*([0-9]+(?:\.[0-9]+)?)x/gi)
-  );
-  if (matches.length === 0) return null;
-  const last = matches[matches.length - 1] as RegExpMatchArray | undefined;
-  const num = last?.[1];
-  return num ? parseFloat(num) : null;
-}
-
-function summarizeFfmpegError(stderr: string): string {
-  const lines = stderr.trim().split(/\r?\n/).filter(Boolean);
-  return lines.length > 0 ? lines[lines.length - 1]! : "unknown error";
-}
-
-async function runFfmpegToNull(
-  ffmpegCmd: string,
-  args: string[]
-): Promise<{ code: number; stderr: string }> {
-  const proc = Bun.spawn({
-    cmd: [ffmpegCmd, ...args],
-    stdout: "ignore",
-    stderr: "pipe",
-  });
-  const [code, stderr] = await Promise.all([proc.exited, proc.stderr?.text()]);
-  return { code, stderr: stderr ?? "" };
-}
-
-async function benchmarkCpu(
-  ffmpegCmd: string,
-  sampleInput: string,
-  durationSec: number,
-  cpuPreset: string,
-  crf: number,
-  encoder: string,
-  fallbackBitrate: string
-): Promise<number> {
-  const args = [
-    "-hide_banner",
-    "-y",
-    "-i",
-    sampleInput,
-    "-t",
-    String(durationSec),
-    "-map",
-    "0:v:0",
-    "-c:v",
-    encoder,
-    ...(encoder === "libx264"
-      ? ["-preset", cpuPreset, "-crf", String(crf)]
-      : ["-b:v", fallbackBitrate]),
-    "-pix_fmt",
-    "yuv420p",
-    "-f",
-    "null",
-    "-",
-  ];
-  const res = await runFfmpegToNull(ffmpegCmd, args);
-  if (res.code !== 0) {
-    throw new Error(
-      `ffmpeg exited with code ${res.code}: ${summarizeFfmpegError(res.stderr)}`
-    );
-  }
-  const speed = parseSpeedFromFfmpegLog(res.stderr);
-  if (speed != null) return speed;
-  // Retry with default verbosity if not found
-  const args2 = [
-    "-hide_banner",
-    "-y",
-    "-i",
-    sampleInput,
-    "-t",
-    String(durationSec),
-    "-map",
-    "0:v:0",
-    "-c:v",
-    encoder,
-    ...(encoder === "libx264"
-      ? ["-preset", cpuPreset, "-crf", String(crf)]
-      : ["-b:v", fallbackBitrate]),
-    "-pix_fmt",
-    "yuv420p",
-    "-f",
-    "null",
-    "-",
-  ];
-  const res2 = await runFfmpegToNull(ffmpegCmd, args2);
-  if (res2.code !== 0) {
-    throw new Error(
-      `ffmpeg exited with code ${res2.code}: ${summarizeFfmpegError(
-        res2.stderr
-      )}`
-    );
-  }
-  return parseSpeedFromFfmpegLog(res2.stderr) ?? 1.0;
-}
-
-async function benchmarkQsv(
-  ffmpegCmd: string,
-  sampleInput: string,
-  durationSec: number,
-  globalQuality: number
-): Promise<number> {
-  const args = [
-    "-hide_banner",
-    "-y",
-    "-hwaccel",
-    "qsv",
-    "-hwaccel_output_format",
-    "qsv",
-    "-i",
-    sampleInput,
-    "-t",
-    String(durationSec),
-    "-map",
-    "0:v:0",
-    "-vf",
-    "scale_qsv=format=nv12",
-    "-c:v",
-    "h264_qsv",
-    "-preset",
-    "medium",
-    "-global_quality",
-    String(globalQuality),
-    "-look_ahead",
-    "1",
-    "-f",
-    "null",
-    "-",
-  ];
-  const res = await runFfmpegToNull(ffmpegCmd, args);
-  if (res.code !== 0) {
-    throw new Error(
-      `ffmpeg exited with code ${res.code}: ${summarizeFfmpegError(res.stderr)}`
-    );
-  }
-  const speed = parseSpeedFromFfmpegLog(res.stderr);
-  if (speed == null) {
-    throw new Error("Failed to parse QSV benchmark speed");
-  }
-  return speed;
-}
-
-async function benchmarkNvenc(
-  ffmpegCmd: string,
-  sampleInput: string,
-  durationSec: number,
-  cq: number
-): Promise<number> {
-  const args = [
-    "-hide_banner",
-    "-y",
-    "-hwaccel",
-    "cuda",
-    "-hwaccel_output_format",
-    "cuda",
-    "-i",
-    sampleInput,
-    "-t",
-    String(durationSec),
-    "-map",
-    "0:v:0",
-    "-vf",
-    "scale_cuda=format=yuv420p",
-    "-c:v",
-    "h264_nvenc",
-    "-preset",
-    "medium",
-    "-cq",
-    String(cq),
-    "-f",
-    "null",
-    "-",
-  ];
-  const res = await runFfmpegToNull(ffmpegCmd, args);
-  if (res.code !== 0) {
-    throw new Error(
-      `ffmpeg exited with code ${res.code}: ${summarizeFfmpegError(res.stderr)}`
-    );
-  }
-  const speed = parseSpeedFromFfmpegLog(res.stderr);
-  if (speed == null) {
-    throw new Error("Failed to parse NVENC benchmark speed");
-  }
-  return speed;
-}
-
-async function benchmarkAmf(
-  ffmpegCmd: string,
-  sampleInput: string,
-  durationSec: number,
-  qpI: number
-): Promise<number> {
-  const args = [
-    "-hide_banner",
-    "-y",
-    "-i",
-    sampleInput,
-    "-t",
-    String(durationSec),
-    "-map",
-    "0:v:0",
-    "-c:v",
-    "h264_amf",
-    "-quality",
-    "balanced",
-    "-rc",
-    "cqp",
-    "-qp_i",
-    String(qpI),
-    "-qp_p",
-    String(qpI + 2),
-    "-qp_b",
-    String(qpI + 4),
-    "-f",
-    "null",
-    "-",
-  ];
-  const res = await runFfmpegToNull(ffmpegCmd, args);
-  if (res.code !== 0) {
-    throw new Error(
-      `ffmpeg exited with code ${res.code}: ${summarizeFfmpegError(res.stderr)}`
-    );
-  }
-  const speed = parseSpeedFromFfmpegLog(res.stderr);
-  if (speed == null) {
-    throw new Error("Failed to parse AMF benchmark speed");
-  }
-  return speed;
-}
-
-async function determineCpuConcurrencyViaScaling(
-  ffmpegCmd: string,
-  sampleInput: string,
-  durationSec: number,
-  preset: string,
-  crf: number,
-  encoder: string,
-  fallbackBitrate: string
-): Promise<number> {
-  const maxToTry = Math.max(
-    1,
-    Math.min(6, Math.floor((os.cpus()?.length ?? 8) / 2))
-  );
-  let bestC = 1;
-  let bestThroughput = 0;
-  for (let c = 1; c <= maxToTry; c++) {
-    const procs: Promise<{ code: number; stderr: string }>[] = [];
-    for (let i = 0; i < c; i++) {
-      procs.push(
-        runFfmpegToNull(ffmpegCmd, [
-          "-hide_banner",
-          "-y",
-          "-i",
-          sampleInput,
-          "-t",
-          String(durationSec),
-          "-map",
-          "0:v:0",
-          "-c:v",
-          encoder,
-          ...(encoder === "libx264"
-            ? ["-preset", preset, "-crf", String(crf)]
-            : ["-b:v", fallbackBitrate]),
-          "-pix_fmt",
-          "yuv420p",
-          "-f",
-          "null",
-          "-",
-        ])
-      );
-    }
-    const start = performance.now();
-    const results = await Promise.all(procs);
-    for (const res of results) {
-      if (res.code !== 0) {
-        throw new Error(
-          `ffmpeg exited with code ${res.code}: ${summarizeFfmpegError(
-            res.stderr
-          )}`
-        );
-      }
-    }
-    const elapsedSec = (performance.now() - start) / 1000;
-    const totalProcessed = c * durationSec;
-    const throughput = totalProcessed / Math.max(0.001, elapsedSec);
-    if (throughput > bestThroughput) {
-      bestThroughput = throughput;
-      bestC = c;
-    } else if (c > 2) {
-      // Diminishing returns, break early
-      break;
-    }
-  }
-  return Math.max(1, bestC);
-}
-
-async function writeConfig(
-  pathStr: string,
-  cfg: BenchmarkConfig
-): Promise<void> {
-  await Bun.write(pathStr, JSON.stringify(cfg, null, 2));
-}
-
-async function readConfig(pathStr: string): Promise<BenchmarkConfig | null> {
-  try {
-    const file = Bun.file(pathStr);
-    const exists = await file.exists();
-    if (!exists) {
-      return null;
-    }
-    const text = await file.text();
-    const parsed = JSON.parse(text) as BenchmarkConfig;
-    return parsed;
-  } catch {
-    return null;
-  }
-}
-
-function padNumber(num: number, width = 3): string {
-  const s = String(Math.max(0, Math.floor(num)));
-  return s.length >= width ? s : "0".repeat(width - s.length) + s;
-}
-
-const VIDEO_EXTENSIONS = new Set([
-  ".mp4",
-  ".mkv",
-  ".avi",
-  ".mov",
-  ".wmv",
-  ".m4v",
-  ".webm",
-  ".ts",
-  ".flv",
-  ".m2ts",
-  ".vob",
-]);
-
-function isVideoFile(p: string): boolean {
-  const ext = path.extname(p).toLowerCase();
-  return VIDEO_EXTENSIONS.has(ext);
-}
-
-interface EpisodeCandidate {
-  inputPath: string;
-  baseName: string;
-  episode: number;
-  relativeDir: string;
-}
-
-interface FolderBatch {
-  relativeDir: string;
-  episodes: EpisodeCandidate[];
-  width: number;
-}
-
-interface ConversionJob {
-  candidate: EpisodeCandidate;
-  outputPath: string;
-  outputRelativePath: string;
-  inputRelativePath: string;
-  outputFileName: string;
-}
-
-function extractEpisodeNumberFromName(name: string): number | null {
-  // Remove file extension first
-  const nameWithoutExt = name.replace(/\.[^.]+$/, "");
-
-  // Remove only YouTube IDs added by yt-dlp (11-character alphanumeric strings in brackets)
-  // This handles patterns like: [-ayFQnecY-4], [dQw4w9WgXcQ], etc.
-  // But preserves other bracketed content like [1080p], [h264], [Season 1], etc.
-  const cleaned = nameWithoutExt.replace(/\[([a-zA-Z0-9_-]{11})\]/g, "").trim();
-
-  const lower = cleaned.toLowerCase();
-  // 1) SxxEyy pattern
-  const se = lower.match(
-    /s(\d{1,2})\s*[^a-z0-9]?\s*e(\d{1,3})/i
-  ) as RegExpMatchArray | null;
-  if (se && se[2]) return parseInt(se[2], 10);
-  // 2) Eyy pattern
-  const eonly = lower.match(
-    /e(?:p|pisode)?[\s_-]*(\d{1,3})/
-  ) as RegExpMatchArray | null;
-  if (eonly && eonly[1]) return parseInt(eonly[1], 10);
-  // 3) Leading numbers
-  const lead = lower.match(/^(\d{1,3})\D/) as RegExpMatchArray | null;
-  if (lead && lead[1]) return parseInt(lead[1], 10);
-  // 4) Any number groups; prefer 1-3 digits and avoid common resolution/codec numbers
-  const forbidden = new Set([
-    "1080",
-    "720",
-    "2160",
-    "480",
-    "360",
-    "264",
-    "265",
-    "10",
-  ]);
-  const nums = Array.from(lower.matchAll(/\b(\d{1,3})\b/g))
-    .map((m) => m[1] ?? "")
-    .filter(Boolean) as string[];
-  const filtered = nums.filter((n) => !forbidden.has(n));
-  if (filtered.length > 0) {
-    // Heuristic: take the last small number
-    const lastNum = filtered[filtered.length - 1] as string;
-    return parseInt(lastNum, 10);
-  }
-  return null;
-}
-
-async function listFilesRecursive(dir: string): Promise<string[]> {
-  const out: string[] = [];
-  try {
-    // Check if directory exists first
-    await fs.access(dir);
-    for await (const entry of new Bun.Glob("**/*").scan({ cwd: dir })) {
-      out.push(path.resolve(dir, entry));
-    }
-  } catch (error) {
-    throw new Error(
-      `Cannot access directory "${dir}": ${
-        error instanceof Error ? error.message : String(error)
-      }`
-    );
-  }
-  return out;
-}
-
-async function collectFolderBatches(inputDir: string): Promise<FolderBatch[]> {
-  try {
-    await fs.access(inputDir);
-  } catch (error) {
-    throw new Error(
-      `Cannot access directory "${inputDir}": ${
-        error instanceof Error ? error.message : String(error)
-      }`
-    );
-  }
-
-  const batches: FolderBatch[] = [];
-
-  const walk = async (
-    currentDir: string,
-    relativeDir: string
-  ): Promise<void> => {
-    const entries = await fs.readdir(currentDir, { withFileTypes: true });
-    entries.sort((a, b) =>
-      a.name.localeCompare(b.name, undefined, {
-        numeric: true,
-        sensitivity: "base",
-      })
-    );
-
-    const episodes: EpisodeCandidate[] = [];
-
-    for (const entry of entries) {
-      const entryPath = path.join(currentDir, entry.name);
-      if (entry.isFile()) {
-        if (!isVideoFile(entry.name)) continue;
-        const episode = extractEpisodeNumberFromName(entry.name);
-        if (episode == null) continue;
-        episodes.push({
-          inputPath: entryPath,
-          baseName: entry.name,
-          episode,
-          relativeDir,
-        });
-        continue;
-      }
-      if (entry.isDirectory()) {
-        const nextRelative = relativeDir
-          ? path.join(relativeDir, entry.name)
-          : entry.name;
-        await walk(entryPath, nextRelative);
-      }
-    }
-
-    if (episodes.length > 0) {
-      episodes.sort((a, b) => a.episode - b.episode);
-      const maxEpisode = episodes[episodes.length - 1]!.episode;
-      batches.push({
-        relativeDir,
-        episodes,
-        width: chooseWidth(maxEpisode),
-      });
-    }
-  };
-
-  await walk(inputDir, "");
-  return batches;
-}
-
-async function buildConversionJobs(
-  inputDir: string,
-  outputDir: string
-): Promise<ConversionJob[]> {
-  const batches = await collectFolderBatches(inputDir);
-  const jobs: ConversionJob[] = [];
-
-  for (const batch of batches) {
-    const targetDir = batch.relativeDir
-      ? path.join(outputDir, batch.relativeDir)
-      : outputDir;
-
-    for (const candidate of batch.episodes) {
-      const outputFileName = `${padNumber(candidate.episode, batch.width)}.mp4`;
-      const outputPath = path.join(targetDir, outputFileName);
-      const inputRelativePath = batch.relativeDir
-        ? path.join(batch.relativeDir, candidate.baseName)
-        : candidate.baseName;
-      const outputRelativePath = batch.relativeDir
-        ? path.join(batch.relativeDir, outputFileName)
-        : outputFileName;
-
-      jobs.push({
-        candidate,
-        outputPath,
-        outputRelativePath,
-        inputRelativePath,
-        outputFileName,
-      });
-    }
-  }
-
-  return jobs;
-}
-
-function chooseWidth(maxEpisode: number): number {
-  if (maxEpisode >= 1000) return 4;
-  return 3;
-}
+import type {
+  BenchmarkConfig,
+  ConvertOptions,
+  EncoderChoice,
+} from "./src/config";
+import { DEFAULT_CONFIG, resolvePath, getDefaultConfigPath } from "./src/config";
+import { ensureFfmpegTools } from "./src/ffmpeg-bootstrap";
+import {
+  getEncoders,
+  supportsQsv,
+  supportsNvenc,
+  supportsAmf,
+  benchmarkCpu,
+  benchmarkQsv,
+  benchmarkNvenc,
+  benchmarkAmf,
+  determineCpuConcurrencyViaScaling,
+  writeConfig,
+  readConfig,
+} from "./src/encoders";
+import {
+  isVideoFile,
+  listFilesRecursive,
+  buildConversionJobs,
+  type ConversionJob,
+} from "./src/scan";
 
 async function convertOne(
   ffmpegCmd: string,
@@ -1179,36 +147,6 @@ async function convertOne(
       await fs.unlink(outputPath);
     } catch {}
     throw new Error(`ffmpeg failed for ${inputPath}`);
-  }
-}
-
-class Semaphore {
-  private slots: number;
-  private waiters: Array<() => void> = [];
-  constructor(initial: number) {
-    this.slots = initial;
-  }
-  tryAcquire(): boolean {
-    if (this.slots > 0) {
-      this.slots--;
-      return true;
-    }
-    return false;
-  }
-  async acquire(): Promise<void> {
-    if (this.slots > 0) {
-      this.slots--;
-      return;
-    }
-    await new Promise<void>((resolve) => this.waiters.push(resolve));
-  }
-  release(): void {
-    if (this.waiters.length > 0) {
-      const fn = this.waiters.shift()!;
-      fn();
-    } else {
-      this.slots++;
-    }
   }
 }
